STREAM-117 Fixed the bug to provide environment variable in ingest-config.
Things were already in place for schema command but it was not working for script command.
After analysing, found that there were two different methods to read YAML. In one method we missed to add the deserializing logic that will read the environment variable, hence removed it.
Now we have a common logic to read YAML file along with environment variable across the application.