STREAM-128 fetch location and fileType info for hive tables
Also re-structured existing code related to source as jdbc type and destination as snowflake.
In the existing coode most of the code was using static methods and never changing fields were passed to methods to access it.
Now made them non-static and made constant fields as class level parameter so that any method can access it.
This also helped to get Jdbc and Connection details in HiveAttributeRetriever.